+++ src/components/AudioPlayer.tsx
@@ -12,10 +12,10 @@ const css = StyleSheet.create({
     position: 'absolute',
     top: 0,
     left: 0,
-    bottom: 0,
-    right: 0,
     width: 0,
     height: 0,
+    opacity: 0,
+    overflow: 'hidden',
   },
 })