+++ src/stores/callStore.ts
@@ -70,7 +70,24 @@ export class CallStore {
     if (!uuid || !n) {
       return
     }
-    const c = this.getCallKeep(uuid)
+    // fix bug ios turn off pn on the server side side
+    let c = this.calls.find(_ => _.bugIosOffPnServer && _.pnId === n.id)
+    if (c) {
+      if (c.callkeepUuid !== uuid) {
+        const del = c.callkeepUuid
+        delete this.callkeepMap[del]
+        delete this.callkeepActionMap[del]
+        delete this.calleeRejectedMap[del]
+        RNCallKeep.endCall(del)
+        c.callkeepUuid = uuid
+        c.bugIosOffPnServer = false
+        if (c.answered) {
+          RNCallKeep.answerIncomingCall(uuid)
+        }
+      }
+    } else {
+      c = this.getCallKeep(uuid)
+    }
     if (n.sipPn.autoAnswer && c) {
       if (RnAppState.foregroundOnce && AppState.currentState !== 'active') {
         RNCallKeep.backToForeground()
@@ -478,12 +495,8 @@ export class CallStore {
       c.callkeepUuid = this.callkeepUuidPending
       this.callkeepUuidPending = ''
     }
-    if (
-      !isWeb &&
-      c.incoming &&
-      !c.callkeepUuid &&
-      !ca?.pushNotificationEnabled
-    ) {
+
+    const displayCall = async () => {
       const uuid = newUuid().toUpperCase()
       c.callkeepUuid = uuid
       const op = {
@@ -505,6 +518,20 @@ export class CallStore {
         undefined,
         op,
       )
+    }
+
+    if (
+      !isWeb &&
+      c.incoming &&
+      !c.callkeepUuid &&
+      !ca?.pushNotificationEnabled
+    ) {
+      await displayCall()
+    }
+    // fix bug ios turn off pn on the server side side
+    if (isIos && c.incoming && !c.callkeepUuid) {
+      await displayCall()
+      c.bugIosOffPnServer = true
     }
     // get and check callkeep if pending incoming call
     if (isWeb || !c.incoming || c.answered) {